Extract composer submission cleanup helpers
